Re-worked the plantuml generator to use the older plugin.
Newer plugin supported links, however the diagram would not be scaled to the page. Will have to look at this issue again later to see if we can use the new plugin and still have the image scale correctly. For the time being that means that the links will not work directly from gitbooks unless the person opens the diagrams in a separate page.