feat: integrate cross border docs (#469)
* Integrate cross border docs

Added Product level documentation to explain interscheme, FX and cross-border transactions, all linking to Paul Baker's more technical documentation. Updated the Use Cases document to integrate the overview documents and create a narrative.

* Fixed two typos

Missing "/" in link

* Clarify the role of the Proxy

Made it clear that the Proxy is not involved in the agreement of terms.