Add an overview of Moja Security (#524)
* Add an overview of Moja Security

Added a Security section to the Product documentation, including a subsection introducing HTLC. Expanded a little on Agreement of Terms (because of its link to HTLC). Added links to the security section in the feature list and in the sidebar menu,

* Updates following review

Updates following peer review. Also updates to package-lock.json to address versioning issues.

* Update security.md

* Updated the HTLC page

Reviewed with Michael Richards.